feat: Rename pages and create form page
