+++ src/pages/index.js
@@ -1,1 +1,2 @@
-export { default as LandingPage } from "./LandingPage/LandingPage";
+export { default as LandingPage } from "./landing/LandingPage";
+export { default as FormPage } from "./apply/FormPage";
